Refactor sentry configuration
In the version I have attempted to tackle the issue of Sentry
configuration from a clean slate. I have removed the code in
rqworker that falls back the the `SENTRY_DSN` Django
setting, on the basis that if the client user has set this in
their project settings.py, then they are capable of configuring
Sentry there.

If the user passes `sentry_dsn` explicitly to the command,
then this will override any configuration in the Django
settings, and replace them with a configuration that includes
the RqIntegration and DjangoIntegration.