Update rqworker command Sentry configuration (alternative version) (#442)
* Update rqworker command to incorporate existing sentry config

I have added a method to the command to combine existing Sentry config
options with the options passed in on the command invocation. This means
that if you have initialised Sentry in the Django settings module with specific
options, these will be maintained.

Fixes #440

* Add test for sentry options

* Add tests for rqworker Command.sentry_options method

* Bump up coverage on tests to get through codecov

* Bump test coverage

* Add coverage to dev-packages

* Refactor previous implementation (WIP)

* Refactor to simplify sentry config

* Fix failing tests

* Remove duplicate integrations key from sentry options

* Refactor sentry configuration

In the version I have attempted to tackle the issue of Sentry
configuration from a clean slate. I have removed the code in
rqworker that falls back the the `SENTRY_DSN` Django
setting, on the basis that if the client user has set this in
their project settings.py, then they are capable of configuring
Sentry there.

If the user passes `sentry_dsn` explicitly to the command,
then this will override any configuration in the Django
settings, and replace them with a configuration that includes
the RqIntegration and DjangoIntegration.

* Add RedisIntegration to Sentry configuration