Handle the case where the Connection to Redis uses Sentinels (#671)
Currently, when PyYAML is not installed, `rqstats -y` prints an error
message, but it still exits with a 0 exit code, indicating success. Fix
that, along with a few related issues:

* The error message was printed to stdout rather than stderr.

* The name of the library in the message was incorrect (LibYAML is a C
  library; the Python library is PyYAML).

After this change, you need to install PyYAML to run the testsuite, so
modify the GitHub workflow accordingly. Also, make sure to install all
packages at once, to make sure that later install commands don't overwrite
the result of the earlier ones.

Co-authored-by: Roman Donchenko <rdonchen@outlook.com>