fix Django admin integration including preserving compatibility (#754)
This change fixes the Django admin integration added in
https://github.com/rq/django-rq/pull/747. As is, the integration adds
the views, but they do not reverse properly without also including the
URLs via `include("django_rq.urls")`. When using Django's `ModelAdmin`,
the URLs registered will be in the "admin" namespace, but the templates
were not updated to include a namespace so they fail to resolve.

This change conditionally resolves with either the prefix `django_rq:`
or `admin:django_rq_` depending on how the view is called (via the admin
app or not) and the URLs are also registered with the corresponding
prefix.

In order to properly test this integration, the default URL conf has been
updated to not include the django_rq URLS via `include`, and a separate
`default_with_custom_mount_urls` URL conf is provided in order to test
integrations using both `include` and the admin integration.